BUG: fix test_buffer_hashlib for 64-bit systems

+++ numpy/core/tests/test_regression.py
@@ -1242,7 +1242,7 @@ class TestRegression(TestCase):
         except ImportError:
             from md5 import new as md5
 
-        x = np.array([1,2,3])
+        x = np.array([1,2,3], dtype=np.int32)
         assert_equal(md5(x).hexdigest(), '2a1dd1e1e59d0a384c26951e316cd7e6')
 
 if __name__ == "__main__":
